+++ packages/integrations/src/propresenter/index.js
@@ -1,0 +1,2 @@
+export { defaultProPresenterConfig } from "./default-config.js";
+export { ProPresenterAdapter } from "./propresenter-adapter.js";